+++ Printf/libft/libft.h
@@ -69,6 +69,6 @@ t_list				*ft_lstmap(t_list *lst, void *(*f)(void *),
 					void (*del)(void *));
 char				*ft_itoa_u(unsigned int n);
 char				*ft_itoa_hex(int n);
-char				*ft_itoa_uhex(unsigned int n);
+char				*ft_itoa_uhex(unsigned long int n);
 char				*ft_strmultiply(char *c, size_t times);
 #endif
